Minor Changes
clusterOutput.c/h: Removed the requirement for the name of the executable from errorMessage()

node.c: Put a temporary check on the number of children spawned. This serves as a way to observe the behavior without fork bombing

uici.c/h: Variable name changes, an extra feedback message, and getting rid of the requirement to provide a host name to u_connect. The name was previous used for user feedback but that's no necessary and too difficult to implement compared to the actual value of the thing.