A Director Program to Connect to the Nodes
For now the only use this file has is to connect to the remote servers. u_connect() has feedback messages. The simplicity of the file is so the behavior of the program can be observed before getting more involved with the message passing or reads and writes.

WIP: Read and write to the nodes.