Documentation and Tweaks
r_read() was documented to better illustrate what exactly is going on. The expected behavior from a function called r_read() would be to read from something and return what was read. With that in mind r_read() was made to return the char* of the <size> bytes taken from the socket. The previous version used a pass by reference method that works but is less intuitive.