Conform Name Resolution Functions to Expected Behavior
The behavior expected by functions nameToAddr() and addrToName() would be to take in a name and address respectively. The return expected would then be an address and a name respectively.

Due to that change the u_accept() function doesn't need to take in a char* hostName or int hostNameSize variable.